+++ R/zzz.R
@@ -36,12 +36,6 @@
   }
 }
 
-.onAttach <- function(libname, pkgname) {
-  packageStartupMessage(
-    sprintf("%s %s", pkgname, utils::packageVersion(pkgname))
-  )
-}
-
 utils::globalVariables(c(
   "badge", "category", "charge_state", "comp_id", "component",
   "composite_score", "count", "display_id", "enzyme", "flag",